[MOVE/#32] 프로필 패키지 분리
