+++ app/src/main/AndroidManifest.xml
@@ -72,15 +72,15 @@
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.main.profile.ProfileManageActivity"
+            android:name=".presentation.main.profile.manage.ProfileManageActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.main.profile.ProfileQuitActivity"
+            android:name=".presentation.main.profile.manage.ProfileQuitActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.main.profile.ProfileQuitForSureActivity"
+            android:name=".presentation.main.profile.manage.ProfileQuitForSureActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
