+++ app/src/main/java/com/yello/presentation/main/MainActivity.kt
@@ -10,8 +10,8 @@ import com.yello.R
 import com.yello.databinding.ActivityMainBinding
 import com.yello.presentation.main.look.LookFragment
 import com.yello.presentation.main.myyello.MyYelloFragment
-import com.yello.presentation.main.profile.ProfileFragment
-import com.yello.presentation.main.profile.ProfileViewModel
+import com.yello.presentation.main.profile.info.ProfileFragment
+import com.yello.presentation.main.profile.info.ProfileViewModel
 import com.yello.presentation.main.recommend.RecommendFragment
 import com.yello.presentation.main.yello.YelloFragment
 import dagger.hilt.android.AndroidEntryPoint
