+++ app/src/main/res/layout/fragment_profile_friend_item_bottom_sheet.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.yello.presentation.main.profile.ProfileViewModel" />
+            type="com.yello.presentation.main.profile.info.ProfileViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
